Add common argument parsing implementation to all scripts
